Use open with explicitly specifying an encoding
The inbuilt function `open` should always get an argument for
`encoding`, otherwise the behaviour of the program will depend on what
OS the code is running on. This is especially problematic for Windows,
as Windows still likes to rely on different encodings that are not
utf-8.

This commit fixes all of the instances detected by the `pylint` error
`unspecified-encoding`.