Samples: Python 3.13, internal visualizer
The Zivid Visualizer is now available in Python and we have updated the
samples to use it instead of Open3D. Since fewer samples require Open3D,
we add support for Python 3.13. The samples that require Open3D will
still warn if Open3D is not installed, and they will for now require
Python 3.12 or earlier.